+++ BlackJack.BusinessLogic/Interfaces/IGameLogicService.cs
@@ -5,19 +5,19 @@ namespace BlackJack.BusinessLogic.Interfaces
 {
     public interface IGameLogicService
     {
-        Task<string> ValidateBet(int bet, int gamePlayerId);
+        Task<string> ValidateBet(int bet, long gamePlayerId);
 
-        Task<GameLogicRoundFirstPhaseResponseView> DoRoundFirstPhase(int bet, int gameId);
+        Task<GameLogicRoundFirstPhaseResponseView> DoRoundFirstPhase(int bet, long gameId);
 
-        Task<GameLogicRoundFirstPhaseResponseView> ResumeGameAfterRoundFirstPhase(int gameId);
+        Task<GameLogicRoundFirstPhaseResponseView> ResumeGameAfterRoundFirstPhase(long gameId);
 
-        Task<GameLogicAddCardToHumanView> AddCardToHuman(int gameId);
+        Task<GameLogicAddCardToHumanView> AddCardToHuman(long gameId);
 
-        Task<GameLogicRoundSecondPhaseResponseView> DoRoundSecondPhase(int gameId, bool blackJackDangerContinueRound = false);
+        Task<GameLogicRoundSecondPhaseResponseView> DoRoundSecondPhase(long gameId, bool blackJackDangerContinueRound = false);
 
-        Task<GameLogicRoundSecondPhaseResponseView> ResumeGameAfterRoundSecondPhase(int gameId);
+        Task<GameLogicRoundSecondPhaseResponseView> ResumeGameAfterRoundSecondPhase(long gameId);
 
-        Task EndRound(int gameId);
+        Task EndRound(long gameId);
 
         Task EndGame(GameLogicEndGameView endGameViewModel);
     }
