+++ BlackJack.DataAccess/Repositories/Interfaces/IGamePlayerRepository.cs
@@ -6,19 +6,19 @@ namespace BlackJack.DataAccess.Repositories.Interfaces
 {
     public interface IGamePlayerRepository
     {
-        Task<int> GetScoreById(int gamePlayerId);
+        Task<int> GetScoreById(long gamePlayerId);
 
-        Task<GamePlayer> GetSpecificPlayerForStartRound(int gameId, int playerType);
+        Task<GamePlayer> GetSpecificPlayerForStartRound(long gameId, int playerType);
 
-        Task<IEnumerable<GamePlayer>> GetSpecificPlayersForStartRound(int gameId, int playerType);
+        Task<IEnumerable<GamePlayer>> GetSpecificPlayersForStartRound(long gameId, int playerType);
 
-        Task<GamePlayer> GetSpecificPlayerWithoutCards(int gameId, int playerType);
+        Task<GamePlayer> GetSpecificPlayerWithoutCards(long gameId, int playerType);
 
-        Task<IEnumerable<GamePlayer>> GetSpecificPlayersWithoutCards(int gameId, int playerType);
+        Task<IEnumerable<GamePlayer>> GetSpecificPlayersWithoutCards(long gameId, int playerType);
 
-        Task<GamePlayer> GetSpecificPlayerWithCards(int gamePlayerId, int playerCard);
+        Task<GamePlayer> GetSpecificPlayerWithCards(long gamePlayerId, int playerCard);
 
-        Task<IEnumerable<GamePlayer>> GetSpecificPlayersWithCards(int gameId, int playerType);
+        Task<IEnumerable<GamePlayer>> GetSpecificPlayersWithCards(long gameId, int playerType);
         
         Task CreateMany(IEnumerable<GamePlayer> gamePlayers);
 
@@ -28,8 +28,8 @@ namespace BlackJack.DataAccess.Repositories.Interfaces
 
         Task UpdateManyAfterRoundSecondPhase(IEnumerable<GamePlayer> gamePlayers);
 
-        Task DeleteBotsWithZeroScore(int gameId);
+        Task DeleteBotsWithZeroScore(long gameId);
 
-        Task DeleteAllByGameId(int gameId);
+        Task DeleteAllByGameId(long gameId);
     }
 }
