+++ BlackJack.UI/Controllers/GameLogicController.cs
@@ -59,7 +59,7 @@ namespace BlackJack.UI.Controllers
         }
 
         [Route("ResumeGameAfterRoundFirstPhase"), HttpGet]
-        public async Task<IHttpActionResult> ResumeGameAfterRoundFirstPhase(int gameId)
+        public async Task<IHttpActionResult> ResumeGameAfterRoundFirstPhase(long gameId)
         {
             try
             {
@@ -75,7 +75,7 @@ namespace BlackJack.UI.Controllers
         }
         
         [Route("AddCardToHuman"), HttpGet]
-        public async Task<IHttpActionResult> AddCardToHuman(int gameId)
+        public async Task<IHttpActionResult> AddCardToHuman(long gameId)
         {
             try
             {
@@ -107,7 +107,7 @@ namespace BlackJack.UI.Controllers
         }
 
         [Route("ResumeGameAfterRoundSecondPhase"), HttpGet]
-        public async Task<IHttpActionResult> ResumeGameAfterRoundSecondPhase(int gameId)
+        public async Task<IHttpActionResult> ResumeGameAfterRoundSecondPhase(long gameId)
         {
             try
             {
@@ -123,7 +123,7 @@ namespace BlackJack.UI.Controllers
         }
 
         [Route("EndRound"), HttpGet]
-        public async Task<IHttpActionResult> EndRound(int gameId)
+        public async Task<IHttpActionResult> EndRound(long gameId)
         {
             try
             {
